+++ backend/src/main/java/ru/hh/school/checkupextension/contest/ContestService.java
@@ -1,18 +1,17 @@
 package ru.hh.school.checkupextension.contest;
 
 import jakarta.inject.Inject;
-import jakarta.ws.rs.NotAuthorizedException;
 
 import org.springframework.transaction.annotation.Transactional;
-import ru.hh.school.checkupextension.core.integration.CheckupInteraction;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestSubmission;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestProblem;
 import ru.hh.school.checkupextension.core.data.dto.contest.ContestSubmissionResult;
 import ru.hh.school.checkupextension.core.data.entity.ProblemEntity;
 import ru.hh.school.checkupextension.core.data.entity.SubmissionEntity;
+import ru.hh.school.checkupextension.core.integration.CheckupInteraction;
 import ru.hh.school.checkupextension.core.repository.Repository;
-import ru.hh.school.checkupextension.utils.exception.ProblemNotFoundException;
-import ru.hh.school.checkupextension.utils.exception.SubmissionNotFoundException;
+import ru.hh.school.checkupextension.utils.exception.core.ProblemNotFoundException;
+import ru.hh.school.checkupextension.utils.exception.core.SubmissionNotFoundException;
 import ru.hh.school.checkupextension.utils.mapper.ProblemMapper;
 import ru.hh.school.checkupextension.utils.mapper.SubmissionMapper;
 
@@ -21,21 +20,19 @@ public class ContestService {
     private final Repository<ProblemEntity> problemRepository;
     private final Repository<SubmissionEntity> submissionRepository;
 
-    private final CheckupInteraction checkupInteraction;
+    private final CheckupInteraction checkupIntegrator;
 
     @Inject
-    public ContestService(CheckupInteraction checkupInteraction,
+    public ContestService(CheckupInteraction checkupIntegrator,
                           Repository<ProblemEntity> problemRepository,
                           Repository<SubmissionEntity> submissionRepository) {
-        this.checkupInteraction = checkupInteraction;
+        this.checkupIntegrator = checkupIntegrator;
         this.problemRepository = problemRepository;
         this.submissionRepository = submissionRepository;
     }
 
     @Transactional
     public ContestProblem getProblem(Long problemId) {
-        if (checkupInteraction.verifyUserToken(""))
-            throw new NotAuthorizedException("");
 
         var problem = problemRepository.getById(problemId).orElseThrow(() -> new ProblemNotFoundException(problemId));
         return ProblemMapper.toContestProblem(problem);
@@ -43,27 +40,27 @@ public class ContestService {
 
     @Transactional
     public ContestSubmission createSubmission(String userToken, ContestSubmission submission) {
-        verifyUser(userToken);
+        var userInfo = checkupIntegrator.getUserInfo(userToken);
+        var problem = problemRepository
+                .getById(submission.problemId)
+                .orElseThrow(() -> new ProblemNotFoundException(submission.problemId));
 
-        var userId = checkupInteraction.getUserId(userToken);
-        var entity = SubmissionMapper.toNewEntity(userId, submission);
+        var entity = SubmissionMapper.toNewEntity(userInfo.userId(), submission);
         var addedEntity = submissionRepository.create(entity);
+        var dto = SubmissionMapper.toContestDto(addedEntity);
 
-        return SubmissionMapper.toContestDto(addedEntity);
+        return dto;
     }
 
     @Transactional
     public ContestSubmission getSubmission(String userToken, long submissionId) {
-        verifyUser(userToken);
-
+        var userInfo = checkupIntegrator.getUserInfo(userToken);
         var submission = getSubmissionById(submissionId);
         return SubmissionMapper.toContestDto(submission);
     }
 
     @Transactional
     public ContestSubmissionResult getSubmissionStatus(String userToken, long submissionId) {
-        verifyUser(userToken);
-
         var submission = getSubmissionById(submissionId);
         return SubmissionMapper.toContestStatusDto(submission);
     }
@@ -72,10 +69,5 @@ public class ContestService {
          return submissionRepository
                 .getById(submissionId)
                 .orElseThrow(() -> new SubmissionNotFoundException(submissionId));
-    }
-
-    private void verifyUser(String userToken) {
-        if (checkupInteraction.verifyUserToken(userToken))
-            throw new NotAuthorizedException("");
     }
 }
