+++ backend/src/main/java/ru/hh/school/checkupextension/core/data/dto/contest/ContestSubmissionResult.java
@@ -4,10 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import ru.hh.school.checkupextension.utils.constant.SubmissionJsonPropertyName;
 
-public class ContestSubmissionResult {
-    public final Long id;
-    public final String status;
-
+public record ContestSubmissionResult(Long id, String status) {
     @JsonCreator
     public ContestSubmissionResult(
             @JsonProperty(SubmissionJsonPropertyName.ID) Long id,
