--- backend/src/main/java/ru/hh/school/checkupextension/utils/exception/AuthorizedException.java
@@ -1,4 +1,0 @@
-package ru.hh.school.checkupextension.utils.exception;
-
-public class AuthorizedException extends RuntimeException {
-}
