--- backend/src/main/java/ru/hh/school/checkupextension/utils/exception/constant/ExceptionConstants.java
@@ -1,5 +1,0 @@
-package ru.hh.school.checkupextension.utils.exception.constant;
-
-public class ExceptionConstants {
-  public static final String VERIFICATION_ID_NOT_FOUND_MESSAGE = "Verification not found with id: %s";
-}
