+++ backend/src/main/java/ru/hh/school/checkupextension/utils/exception/core/SubmissionNotFoundException.java
@@ -1,0 +1,17 @@
+package ru.hh.school.checkupextension.utils.exception.core;
+
+public class SubmissionNotFoundException extends RuntimeException {
+    static final String SUBMISSION_ID_NOT_FOUND_MESSAGE = "Submission not found with id: %s";
+
+    public SubmissionNotFoundException(String message) {
+        super(message);
+    }
+
+    public SubmissionNotFoundException(long submissionId) {
+        super(String.format(SUBMISSION_ID_NOT_FOUND_MESSAGE, submissionId));
+    }
+
+    public SubmissionNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
