+++ backend/src/main/java/ru/hh/school/checkupextension/utils/exception/integration/AuthorizedException.java
@@ -1,0 +1,4 @@
+package ru.hh.school.checkupextension.utils.exception.integration;
+
+public class AuthorizedException extends RuntimeException {
+}
