+++ backend/src/main/java/ru/hh/school/checkupextension/utils/exception/integration/ExpiredUserTokenException.java
@@ -1,4 +1,4 @@
-package ru.hh.school.checkupextension.utils.exception;
+package ru.hh.school.checkupextension.utils.exception.integration;
 
 public class ExpiredUserTokenException extends RuntimeException {
 }
