+++ backend/src/main/java/ru/hh/school/checkupextension/utils/exception/mapper/integration/AuthorizedExceptionMapper.java
@@ -1,10 +1,10 @@
-package ru.hh.school.checkupextension.utils.exception.mapper;
+package ru.hh.school.checkupextension.utils.exception.mapper.integration;
 
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.ExceptionMapper;
 import jakarta.ws.rs.ext.Provider;
 
-import ru.hh.school.checkupextension.utils.exception.AuthorizedException;
+import ru.hh.school.checkupextension.utils.exception.integration.AuthorizedException;
 
 @Provider
 public class AuthorizedExceptionMapper implements ExceptionMapper<AuthorizedException> {
