--- backend/src/main/java/ru/hh/school/checkupextension/utils/mapper/Mapper.java
@@ -1,2 +1,0 @@
-package ru.hh.school.checkupextension.utils.mapper;public class Mapper {
-}
